refactor: utility permission fx to user_access
